requirements.txt
djangotoolbox
django
argparse
defusedxml
django-cors-headers
django-filter
djangorestframework
djangorestframework-jwt
djangorestframework-xml
wsgiref==0.1.2
pymongo==2.8